feat!: add routing for multiple pages
