+++ package.json
@@ -12,6 +12,7 @@
     "react-bootstrap": "^2.7.4",
     "react-dom": "^18.2.0",
     "react-icons": "^4.8.0",
+    "react-router-dom": "^6.11.2",
     "react-scripts": "5.0.1",
     "reactstrap": "^9.1.9",
     "web-vitals": "^2.1.4"
