+++ src/Components/buttons.css
@@ -131,6 +131,7 @@
     left: 0;
     top: 0;
     border-radius: 10px;
+    z-index: 2;
 }
 
 /* Posiciona o ícone centralizado dentro do botão */
@@ -138,7 +139,7 @@
     position: absolute;
     transform: translate(-50%, -50%);
     color: white;
-    z-index: 1;
+    z-index: 3;
 }
 
 
@@ -151,7 +152,7 @@
 
 /* Estiliza a dica de ferramenta do ícone */
 .navBarHover i[title]::before {
-    background-color: #000;
+    background-color: #1a1a1a;
     color: #fff;
     transition: opacity 0.2s, visibility 0.2s;
 }
