+++ src/index.js
@@ -3,6 +3,7 @@ import ReactDOM from 'react-dom';
 import './index.css'; // Importa o arquivo CSS principal
 import 'bootstrap/dist/css/bootstrap.min.css';
 import App from './App';
+import { BrowserRouter } from 'react-router-dom';
 
 
 ReactDOM.render(
