+++ src/pages/Home/Home.js
@@ -1,0 +1,7 @@
+export default function Home() {
+    return (
+        <div>
+            <h1>PAGINA 1</h1>
+        </div>
+    )
+}
