+++ src/pages/Reports/Reports.js
@@ -1,0 +1,7 @@
+export default function Reports() {
+    return (
+        <div>
+            <h1>PAGINA 2</h1>
+        </div>
+    )
+}
